// src/app/modules/student/student.interface.ts
export type Guardian = {
    fatherName: string;
    fatherOccupation: string;
    fatherContactNo: string;
    MotherName: string;
    MotherOccupation: string;
    MotherContactNo: string;
}


export type UserName  = {
    firstName: string;
    middleName: string;
    lastName: string;
}


export type LocalGuardian = {
    name: string;
    occupation: string;
    contactNo: string;
    address: string;

}



export type Student = {
    id: string;
  name: UserName;
  gender: "male" | "female";
  dateOfBirth?: string;
  email: string;
  contactNo: string;
  emergencyContactNumber: string;
  bloodGroup? : 'A+' | 'A-' | 'B+' | 'B-' | 'AB+' | 'AB-' | 'O+' | 'O-';
  presentAddress: string;
  permanentAddress: string;
  guardian: Guardian ; 
  localGuardian : LocalGuardian ;
  profileImg ? : string;
  isActive : "active" |"blocked"
};